requirements.txt
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-huggingface==0.0.3
langchain-chroma==0.1.3
langchain-groq==0.1.9
unstructured==0.15.0
unstructured[pdf]==0.15.0
nltk==3.8.1
python-magic==0.4.27
chromadb==0.5.17
psycopg_binary==3.2.3